requirements.txt
discord.py==2.3.2
openai==1.51.0
playwright==1.48.0
python-dotenv==1.0.0
beautifulsoup4==4.12.3
aiohttp==3.10.5

